requirements.txt
streamlit>=1.18.0
pymongo>=4.5.0
python-dotenv>=0.19.0
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.0.0

